feat: 이벤트 기반 좋아요 처리와 집계 분리
