--- apps/commerce-api/src/main/java/com/loopers/application/event/LikeEventHandler.java
@@ -1,18 +1,0 @@
-package com.loopers.application.event;
-
-import com.loopers.domain.product.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
-
-@Component
-@RequiredArgsConstructor
-public class LikeEventHandler {
-    private final ProductRepository productRepository;
-
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public void onLikeCreated(LikeCreateEvent event) {
-
-    }
-}
