+++ apps/commerce-api/src/main/java/com/loopers/application/like/LikeCreateEvent.java
@@ -1,0 +1,4 @@
+package com.loopers.application.like;
+
+public record LikeCreateEvent(Long userId, Long productId) {
+}
